backend/requirements.txt
# Core
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy[asyncio]>=2.0.36,<3.0.0
aiosqlite>=0.20.0,<1.0.0
alembic>=1.14.0,<2.0.0
pydantic-settings>=2.6.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0

# Dev
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
httpx>=0.28.0,<1.0.0
